requirements.txt
streamlit
pandas
pinecone
openai
sentence-transformers
google-genai
tqdm
python-dotenv
torch
huggingface-hub
numpy
anthropic